:recycle: Refactor av utils dir
