+++ aksel.nav.no/website/components/hooks/useCheckAuth.ts
