+++ aksel.nav.no/website/components/hooks/useFormatedDate.ts
@@ -1,5 +1,5 @@
-import { dateStr } from "components/website-modules/utils/date-string";
 import { useEffect, useState } from "react";
+import { dateStr } from "@/utils";
 
 export const useFormatedDate = (date: string) => {
   const [_date, setDate] = useState(null);
@@ -8,5 +8,6 @@ export const useFormatedDate = (date: string) => {
     const handleDate = async () => setDate(await dateStr(date));
     handleDate();
   }, [date]);
+
   return _date;
 };
