+++ aksel.nav.no/website/components/logging/useAmplitude.ts
