--- aksel.nav.no/website/components/types/feedback.ts
@@ -1,13 +1,0 @@
-export enum HelpfulArticleEnum {
-  "JA" = "ja",
-  "DELVIS" = "delvis",
-  "NEI" = "nei",
-  "MISC" = "misc",
-}
-
-export type FeedbackT = {
-  message: string;
-  url: string;
-  docId: string;
-  type?: "ja" | "nei" | "forslag" | "footer" | "uu";
-};
