+++ aksel.nav.no/website/components/types/index.ts
@@ -1,4 +1,3 @@
 export * from "./sanity-schema";
 export * from "./next";
-export * from "./feedback";
 export * from "./search";
