+++ aksel.nav.no/website/components/utils/get-authors.ts
@@ -1,0 +1,5 @@
+import { AkselBloggDocT, ResolveContributorsT, ResolveSlugT } from "@/types";
+
+export const getAuthors = (
+  blog: ResolveContributorsT<ResolveSlugT<AkselBloggDocT>>
+) => (blog?.contributors as any)?.map((x) => x?.title) ?? [];
