+++ aksel.nav.no/website/components/utils/get-image.ts
