+++ aksel.nav.no/website/components/utils/is-new.ts
@@ -1,0 +1,7 @@
+export const isNew = (date: string) => {
+  const date1 = new Date(date);
+  const now = new Date();
+  const timeDiff = Math.abs(now.getTime() - date1.getTime());
+  const diffDays = Math.ceil(timeDiff / (1000 * 3600 * 24));
+  return diffDays <= 90;
+};
