+++ aksel.nav.no/website/components/utils/prefers-reduced-motion.ts
@@ -1,4 +1,4 @@
-export function PrefersReducedMotion() {
+export function userPrefersReducedMotion() {
   const QUERY = "(prefers-reduced-motion: no-preference)";
   const mediaQueryList = window.matchMedia(QUERY);
   const prefersReducedMotion = !mediaQueryList.matches;
