+++ aksel.nav.no/website/components/utils/remove-emojies.ts
@@ -1,0 +1,5 @@
+export const removeEmojies = (s: string) =>
+  s.replace(
+    /([\u2700-\u27BF]|[\uE000-\uF8FF]|\uD83C[\uDC00-\uDFFF]|\uD83D[\uDC00-\uDFFF]|[\u2011-\u26FF]|\uD83E[\uDD10-\uDDFF])/g,
+    ""
+  );
