+++ aksel.nav.no/website/components/website-modules/icon-page/Page.tsx
@@ -1,10 +1,10 @@
-import { useMedia } from "@/utils";
+import { useMedia } from "@/hooks/useMedia";
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
 import * as Icons from "@navikt/aksel-icons";
 import meta from "@navikt/aksel-icons/metadata";
 import { HGrid, Heading, Modal, Search, ToggleGroup } from "@navikt/ds-react";
 import cl from "clsx";
-import Footer from "components/layout/footer/Footer";
-import { Header } from "components/layout/header/Header";
 import { SEO } from "components/website-modules/seo/SEO";
 import { SuggestionBlock } from "components/website-modules/suggestionblock/SuggestionBlock";
 import Fuse from "fuse.js";
