--- aksel.nav.no/website/components/website-modules/utils/contexts/id-context.ts
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const IdContext = createContext<{ id?: string }>(null);
