--- aksel.nav.no/website/components/website-modules/utils/contexts/index.tsx
@@ -1,1 +1,0 @@
-export * from "./id-context";
