+++ aksel.nav.no/website/pages/404.tsx
@@ -1,3 +1,5 @@
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
 import { AmplitudeEvents, amplitude } from "@/logging";
 import { BugIcon } from "@navikt/aksel-icons";
 import {
@@ -10,8 +12,6 @@ import {
   Page,
   VStack,
 } from "@navikt/ds-react";
-import Footer from "components/layout/footer/Footer";
-import { Header } from "components/layout/header/Header";
 import { useEffect } from "react";
 
 export default function NotFound() {
