+++ aksel.nav.no/website/pages/_app.tsx
@@ -1,14 +1,14 @@
 import { BaseSEO } from "components/website-modules/seo/BaseSEO";
-import useAmplitudeInit from "components/website-modules/utils/logging/useAmplitude";
-import { useScrollToHashOnPageLoad } from "components/website-modules/utils/util";
 import { AppProps } from "next/app";
 import { useEffect } from "react";
 import { hotjar } from "react-hotjar";
-import { IdContext } from "components/website-modules/utils/contexts/id-context";
+import { SanityDocIdContext } from "@/hooks/useSanityDocId";
+import { useHashScroll } from "@/hooks/useHashScroll";
+import { useAmplitudeInit } from "@/logging";
 import "components/styles/index.css";
 
 function App({ Component, pageProps, router }: AppProps) {
-  useScrollToHashOnPageLoad();
+  useHashScroll();
   useAmplitudeInit();
 
   useEffect(() => {
@@ -21,9 +21,11 @@ function App({ Component, pageProps, router }: AppProps) {
   return (
     <>
       <BaseSEO path={router.asPath} />
-      <IdContext.Provider value={{ id: pageProps?.id ?? pageProps?.page?._id }}>
+      <SanityDocIdContext.Provider
+        value={{ id: pageProps?.id ?? pageProps?.page?._id }}
+      >
         <Component {...pageProps} />
-      </IdContext.Provider>
+      </SanityDocIdContext.Provider>
     </>
   );
 }
