+++ aksel.nav.no/website/pages/god-praksis/artikler/index.tsx
@@ -1,3 +1,5 @@
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
 import { getClient } from "@/sanity/client.server";
 import { contributorsSingle } from "@/sanity/queries";
 import {
@@ -8,8 +10,6 @@ import {
   ResolveTemaT,
 } from "@/types";
 import { Button, ErrorMessage, Heading } from "@navikt/ds-react";
-import Footer from "components/layout/footer/Footer";
-import { Header } from "components/layout/header/Header";
 import ArtikkelCard from "components/sanity-modules/cards/ArtikkelCard";
 import { AkselCubeStatic } from "components/website-modules/aksel-cube/AkselCube";
 import { SEO } from "components/website-modules/seo/SEO";
