+++ aksel.nav.no/website/pages/index.tsx
@@ -1,7 +1,9 @@
-import Footer from "components/layout/footer/Footer";
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
 import { getClient } from "@/sanity/client.server";
 import { contributorsAll } from "@/sanity/queries";
 import { AkselTemaT, NextPageT } from "@/types";
+import { userPrefersReducedMotion } from "@/utils";
 import {
   CompassIcon,
   ComponentIcon,
@@ -11,7 +13,6 @@ import {
 } from "@navikt/aksel-icons";
 import { Heading, Ingress } from "@navikt/ds-react";
 import cl from "clsx";
-import { Header } from "components/layout/header/Header";
 import GodPraksisCardSimple from "components/sanity-modules/cards/GodPraksisCardSimple";
 import FrontpageBlock, {
   BlocksT,
@@ -19,7 +20,6 @@ import FrontpageBlock, {
 import { IntroCards } from "components/website-modules/IntroCards";
 import { AkselCubeAnimated } from "components/website-modules/aksel-cube/AkselCube";
 import { SEO } from "components/website-modules/seo/SEO";
-import { PrefersReducedMotion } from "components/website-modules/utils/prefers-reduced-motion";
 import { GetStaticProps } from "next/types";
 import { Suspense, lazy, useEffect, useState } from "react";
 
@@ -124,7 +124,7 @@ const Forside = ({ page, tema, blocks }: PageProps["props"]) => {
       navigator.userAgent.indexOf("Safari") !== -1 &&
       navigator.userAgent.indexOf("Chrome") === -1;
 
-    setReducedMotion(PrefersReducedMotion() || disableAnimations);
+    setReducedMotion(userPrefersReducedMotion() || disableAnimations);
     const data = localStorage.getItem("pause-animations");
     if (disableAnimations) {
       setPause(true);
