+++ aksel.nav.no/website/pages/komponenter/[...slug].tsx
@@ -1,3 +1,6 @@
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
+import { WithSidebar } from "@/layout/templates/WithSidebar";
 import { amplitude, AmplitudeEvents } from "@/logging";
 import { SanityBlockContent } from "@/sanity-block";
 import { getClient } from "@/sanity/client.server";
@@ -19,9 +22,6 @@ import {
   GithubIcon,
   YarnIcon,
 } from "components/assets/Icons";
-import Footer from "components/layout/footer/Footer";
-import { Header } from "components/layout/header/Header";
-import { WithSidebar } from "components/layout/templates/WithSidebar";
 import ComponentOverview from "components/sanity-modules/component-overview/ComponentOverview";
 import IntroSeksjon from "components/sanity-modules/intro-seksjon/IntroSeksjon";
 import { SEO } from "components/website-modules/seo/SEO";
