+++ aksel.nav.no/website/pages/produktbloggen/[slug].tsx
@@ -1,3 +1,5 @@
+import Footer from "@/layout/footer/Footer";
+import { Header } from "@/layout/header/Header";
 import { SanityBlockContent } from "@/sanity-block";
 import { getClient } from "@/sanity/client.server";
 import { urlFor } from "@/sanity/interface";
@@ -8,16 +10,13 @@ import {
   ResolveContributorsT,
   ResolveSlugT,
 } from "@/types";
-import { abbrName, dateStr } from "@/utils";
+import { abbrName, dateStr, getImage } from "@/utils";
 import { BodyShort, Detail, Heading, Ingress } from "@navikt/ds-react";
-import Footer from "components/layout/footer/Footer";
-import { Header } from "components/layout/header/Header";
 import BloggCard from "components/sanity-modules/cards/BloggCard";
 import { BloggAd } from "components/website-modules/BloggAd";
 import { AkselCubeStatic } from "components/website-modules/aksel-cube/AkselCube";
 import { SEO } from "components/website-modules/seo/SEO";
 import TableOfContents from "components/website-modules/toc/TOC";
-import { getImage } from "components/website-modules/utils/get-image";
 import Image from "next/legacy/image";
 import { GetServerSideProps } from "next/types";
 import { Suspense, lazy } from "react";
